requirements.txt
streamlit
crewai==0.14.0
python-dotenv
playwright
html2text
google-generativeai>=0.3.2
langchain>=0.1.0
langchain-google-genai==2.1.1
google-ai-generativelanguage>=0.6.16
langchain-ollama
